Cleanup, update setup py
